requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.1.0
billiard==4.2.1
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==5.0.6
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-crum==0.7.9
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-phonenumber-field==8.0.0
django-templated-mail==1.1.1
django-timezone-field==7.0
django-tree-queries==0.19.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
flake8==7.1.1
gunicorn==23.0.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.2
phonenumbers==8.13.42
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
PyYAML==6.0.1
redis==5.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.19.1
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
